Fixed testBrickKinematics. collisionDetect now turns taylorvars into double before calling forwardKin (so mex will be on). Minor fix to contactPositions for taylorvars. Added a comment line to doKinematics.

+++ systems/plants/@RigidBodyManipulator/collisionDetect.m
@@ -41,7 +41,7 @@ if ~isstruct(kinsol)
 end
 
 if (kinsol.mex ~= true) 
-  doKinematics(obj,kinsol.q);
+  doKinematics(obj,double(kinsol.q));
   warning('RigidBodyManipulator:collisionDetect:doKinematicsMex', ...
     'Calling doKinematics using mex before proceeding');
 end
